+++ R/autodb.r
@@ -4,9 +4,10 @@
 #' \code{\link{flatten}}, \code{\link{normalise}},
 #' \code{\link{cross_reference}}, and \code{\link{decompose}}, in order.
 #'
+#' Since `decompose` only works with functional dependencies, not approximate
+#' dependencies, the accuracy in `dfd` is fixed as 1.
+#'
 #' @param df a data.frame, containing the data to be normalised.
-#' @param accuracy a numeric in (0, 1], giving the accuracy threshold threshold
-#'   required in order to conclude a dependency.
 #' @param name a scalar character, giving the name of the database. This name
 #'   is used for the resulting graph when using \code{\link{gv.database}}, to
 #'   allow for easier combining of graphs into a single diagram if required.
@@ -34,11 +35,10 @@
 #'   \code{\link{cross_reference}} for details.
 #' @examples
 #' # simple example
-#' autodb(ChickWeight, 1)
+#' autodb(ChickWeight)
 #' @export
 autodb <- function(
   df,
-  accuracy,
   name = NA_character_,
   ensure_lossless = TRUE,
   remove_avoidable = FALSE,
@@ -49,7 +49,7 @@ autodb <- function(
 ) {
   report <- reporter(progress, progress_file)
 
-  dfd(df, accuracy, progress = progress, progress_file = "", ...) |>
+  dfd(df, 1, progress = progress, progress_file = "", ...) |>
     report$op(flatten, "flattening") |>
     report$op(normalise, "normalising", remove_avoidable, constants_name) |>
     report$op(cross_reference, "cross-referencing", ensure_lossless) |>