+++ R/plot.r
@@ -56,7 +56,7 @@
 #' cat(txt_df)
 #' DiagrammeR::grViz(txt_df)
 #' # simple database example
-#' db <- autodb(ChickWeight, 1, "chick")
+#' db <- autodb(ChickWeight, "chick")
 #' txt_db <- gv(db)
 #' cat(txt_db)
 #' DiagrammeR::grViz(txt_db)